backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
speechrecognition==3.10.0
pyttsx3==2.90
langgraph==0.0.20
langchain-openai==0.0.2
langchain-core==0.1.10
python-dotenv==1.0.0
websockets==12.0
pydub==0.25.1